I2C 10-bit + SD logger + exception-policy tests + hygiene
- I2C 10-bit: OAR1 ADDMODE+ADD[9:0] match (no 7-bit aliasing),
  master headers NACK (reserved range, no peers); inject addr u16;
  6 asserts, test_all 532. COVERAGE I2C row closed.
- SD data-logger demo: register-level SDIO + per-RTC-second ADC
  temp, CMD24 log + CMD17 verify (log N rtc=R adc=A ok); ADC shows
  the RC charging curve to nominal. 8/8 + preset + browser + CI.
  Page preset ships a blank 1MiB SDHC image.
- Exception-policy tests (core_tests 10/10): same-prio non-nesting,
  priority-ordered dispatch, bad EXC_RETURN faults, SysTick debt
  re-pend-once (pins the earlier fix natively). No Unicorn oracle:
  generic ARM lacks M-profile stacking (bring-up probed, dropped) —
  noted in fuzz_diff.py.
- Hygiene: dead stretch_until removed (+ misleading doc), AGENTS
  date/counts, NVIC coalescing guidance, i2cInject in USAGE.
- Verified: test_all 532, canary 39/39, cli 200M 39/39, emu.js 3/3,
  cpu 48/48, census both, fuzz canary 0 divs, coremark 5/5,
  formats/esm/ws/bus-tap + all suites, browser 15/15 local
  (3 gh-pages need external network).

+++ src/lib.rs
@@ -490,7 +490,7 @@ pub fn i2c_inject_rx(channel: u8, bytes: &[u8]) {
 /// slave-RX, RXNE unread or ACK cleared; read returns -1 when not in
 /// slave-TX or DR empty (stretch equivalents).
 #[wasm_bindgen]
-pub fn i2c_inject_start(channel: u8, addr: u8, is_read: bool) -> bool {
+pub fn i2c_inject_start(channel: u8, addr: u16, is_read: bool) -> bool {
     match try_sys() {
         Some(sys) => sys.p.i2c_inject_start(sys, channel as u32, addr, is_read),
         None => false,
